Did some huge refactoring regarding input handling
As written above. It needs to be checked, though: I'm not wholly satisfied with the keyboardInput() functions.

The one call in displayer, basically retrieves m_window from Renderer, and then it passes it to every other class in the call chain. I'm not sure if it's supposed to be ok or not...